dev: Big main() updated
This is the start of a series of huge commits.
Basically, I've been sitting on this code base until it was mostly
functional.
Now, not only will it survive all the way through Thriplash, it will
also respond appropriately when a new game with started with the same
players.

As of right now, JITB doesn't actually *answer* the Thriplash, but this
is close enough...
At least close enough to commit these changes.

Changes to be committed:
	modified:   jitb/jitb_main.py

Untracked files:
	jitb/jitb_args.py
	jitb/jitb_globals.py
	jitb/jitb_website.py