JITB-1: Troubleshooting context issues with JitbAi
ANSWERED An original theme for an escape room with Escape from the World's Worst Dad Jokes!

OPENAI'S ANSWER WAS "An original theme for an escape room: 'Escape from the World's Worst Dad Jokes!'"

The An original theme for an escape room: 'Escape from the World's Worst Dad Jokes!' was longer than 45 so it got truncated
ANSWERED Oh dear, now Lin-Manuel Miranda is writing a hip-hop musical about _______ with An original theme for an escape room: 'Escape!

Changes to be committed:
	modified:   jitb/jitb_openai.py