JITB-1: Changed JitbAi OpenAI endpoint
Rolled back the chat.completions endpoint (with gpt-3.5-turbo) to
(temp?) use the old completions endpoing (with gpt-3.5-turbo-instruct)
in and attempt to get rid of this unnecessary pollution of answers with
previous context.

(Maybe there's something I could be doing with chat.completions
assistants but I haven't quite figured it out just yet).

Changes to be committed:
	modified:   jitb/jitb_openai.py

Changes not staged for commit:
	modified:   jitb/jitb_website.py