JITB-1: Polished JitbAi
- Made pycodestyle happy
- Found a better(?) Thriplash instruction (but maybe my sample set is
  too small)

In this situation, I'd normally throw answer_thriplash() through the
unit_test wringer but money is on the line.  But maybe I just do it a
few times with some of the Thriplash prompts I have from my manual logs.

TO DO: DON'T DO NOW...
- A little manual testing of answer_thriplash() (but just a little)
- Pylint polish
- Last looks
- Merge into dev

Changes to be committed:
	modified:   jitb/jitb_openai.py

Untracked files:
	actual_integration.txt