dev: Added a new sub-package jbgames
Extricated functionality into an abstract base class (ABC) so I could
inherit game-specific classes and define game-specific functionality.

Full refactor was functionally tested against two full rounds of
Quiplash 3 in the same room code so it seems the refactor was
successful.

Changes to be committed:
	new file:   jitb/jbgames/jbg_abc.py
	new file:   jitb/jbgames/jbg_page_ids.py
	new file:   jitb/jbgames/jbg_q3.py

Changes not staged for commit:
	modified:   jitb/jitb_globals.py
	modified:   jitb/jitb_website.py