JITB-13: JbgJb module initial commit
Jackbox Games Joke Boat jitb module initial commit.
JbgJb.id_page() is currently passing all 42 unit tests.
The rest of the the class is a copy/paste from Quiplash 2 with most of
the proper names fixed up... but none of the functionality is working
beyond id_page().

TO DO: DON'T DO NOW...
	Write unit tests for some of the get_*() methods of the JbgAbc.
	Implement those methods.

Changes to be committed:
	new file:   jitb/jbgames/jbg_jb.py