JITB-13: Normalized encoding for prompts & answers
Someone has been slipping in non-standard punctuation.
Also, a newline between the end of a prompt and the trailing
fill-in-the-blank really threw polish_answer() off.

Once I polished the prompt, it broke a couple test cases.
Then I had to polish the answer as well.

GOOD NEWS: The openai unit tests are all passing
BAD NEWS: jitb_openai is flush with DEBUGGING noise

Changes to be committed:
	new file:   jitb/jitb_misc.py
	modified:   jitb/jitb_openai.py

Changes not staged for commit:
	modified:   README.md