JITB-13: Fixed(?) Joke Topic generation
I had to work around JitbAi, instead of it working for me, but the
results look fair.

This log file represents one whole game from start to finish.  There's
one answer in there that exceeded the limit (even though the length
limit is explicitly stated to OpenAI) but other than that it was a
flawless run (ignoring the fact JITB lost).

TO DO: DON'T DO NOW:
	- Make the linters happy
	- Live functional testing
	- Merge
	- Move on

Changes to be committed:
	modified:   jitb/jbgames/jbg_jb.py
	new file:   logs/jitb_20240308_225155-0-Working_well.log