JITB-11: Still tweaking new API Joke Topic gen
Still tweaking the Joke Topic generation for the new API.
I think I have the prompt licked.  When it comes to Joke Topic
generation, I had to disable the standard messages and go "raw" for the
AI generation.

It seems to be working in manual unit testing.

TO DO: DON'T DO NOW...
- Add a JitbAi interface to generate multiple promts at once
- Write a JbgJb method to generate multiple prompts that generate 10 answers for everything
- Functionally test refactor
- Remove stale code
- Linters
- Live testing

EOD Commit

Changes to be committed:
	modified:   jitb/jbgames/jbg_jb.py
	modified:   jitb/jitb_openai.py