JITB-42: Refactored join_room() for manual mode
I put a shunt around join_room()'s efforts to completely login if
username is None.  All of that gets bypassed, the function returns, and
the JbgAbc-derived classes see login pages.  Once the user completes the
process, JITB will take over.

As of this commit, all of this is working.  I manually tested manual
mode, pun(?) intended, on Q3 and it worked.

Changes to be committed:
	modified:   jitb/jitb_website.py

Changes not staged for commit:
	modified:   jitb/jitb_main.py

Untracked files:
	test/test_input/JackboxTV-Q3-Login-Twitch_required-error.html
	test/test_input/JackboxTV-Q3-Login-Twitch_required-error_files/
	test/test_input/JackboxTV-Q3-Login-Twitch_required-login.html
	test/test_input/JackboxTV-Q3-Login-Twitch_required-login_files/